1) I've compared Matteo's baseline version (Xoopic_APL_Parallel) to my final M.Sc. version (Xoopic_10.2 on my PC) and apparently they are identical. That's good news because now we can refer to it as a common baseline.
2) I have simulated several validation scenarios on that baseline code
version. The conclusions are:

I've checked again Matteo's final version just to make sure - indeed I
think he has a bug in sptlrgn.cpp, where he conduct the particle advance
twice (one serially and one with my parallel code part)

I've conducted for the baseline version (Xoopic_APL_parallel) a
validation scenario for the DADI field solver in the XY coordinates
(Geometry = 1). Apparently this results in a wrong solution when you use
the DADI field solver with periodic boundary in the y direction
(ElectrostaticFlag = 5 & PeriodicFlagX2 = 1 in my code).

Therefore - when using Cartesian geometry with periodic boundary
condition in the y direction, you have to change the number of cores to
1 (xg/main.cpp). Otherwise - the solution is garbage! . In order to fix
that in the future, i'm pretty sure that the file dadixyp.cpp should be
edited to support the parallelization. Moreover, right now the field
solver in the xy direction (dadixy.cpp) is not parallelized as well.
Although it still gives the right solution, this might as well be
corrected in the future.
Single particle motion looks OK

Also copied Matteo's otools/history.cpp in hope it would fix the h5 file
time history saving.